pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tessera"
version = "0.0.1"
description = "Signed provenance labels and taint-tracking policy for LLM context windows."
requires-python = ">=3.12"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
dependencies = [
    "fastapi>=0.110",
    "httpx>=0.27",
    "pydantic>=2.5",
    "uvicorn>=0.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "pyjwt[crypto]>=2.8",
    "mcp>=1.0",
    "grpcio>=1.70",
    "grpcio-tools>=1.70",
    "protobuf>=4.25",
]
otel = [
    "opentelemetry-api>=1.25",
]
spiffe = [
    "pyjwt[crypto]>=2.8",
    "spiffe>=0.2.5",
]
mcp = [
    "mcp>=1.0",
]
cel = [
    "cel-python>=0.1",
]
sessions = [
    "cryptography>=42.0",
]
agentmesh = [
    "pyyaml>=6.0",
]
xds = [
    "grpcio>=1.70",
    "protobuf>=4.25",
]
langchain = [
    "langchain-core>=0.2",
]
openai-agents = [
    "openai-agents>=0.0.3",
]

[project.scripts]
tessera = "tessera.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
